refector : view 파일 분리
