+++ src/js/view/main.js
@@ -1,9 +1,4 @@
-/* eslint-disable arrow-body-style */
 import { $, $$ } from '../utils/selector.js';
-
-export const showTrialForm = () => {
-  $('.trial-form').classList.remove('hide');
-};
 
 export const hideTrialForm = () => {
   $('.trial-form').classList.add('hide');
@@ -35,16 +30,12 @@ export const toggleDisabledTrial = () => {
   $('.trial-input').disabled = !$('.trial-input').disabled;
 };
 
-export const showDistance = () => {
-  $('.game-result').classList.remove('hide');
-};
-
 export const hideDistance = () => {
   $('.game-result').classList.add('hide');
 };
 
 export const updateDistance = result => {
-  $('.game-result').innerHTML = ` 
+  $('.game-result').innerHTML = `
 		<div class="mt-4 d-flex">
     ${result
       .map(
@@ -58,62 +49,6 @@ export const updateDistance = result => {
   </div>`;
 };
 
-export const showWinner = () => {
-  $('.winner-section').classList.remove('hide');
-};
-
-const alertCongrats = () => {
-  setTimeout(() => alert('축하합니다🎉🎉🎉'), 2000);
-};
-
-export const addLoadingHTML = () => {
-  const loadingHTML = `
-  <div class="d-flex justify-center mt-3 loading">
-    <div class="relative spinner-container">
-      <span class="material spinner"></span>
-    </div>
-  </div> 
-`;
-
-  document.querySelectorAll('.result-container').forEach($el => {
-    const $container = $el.querySelector('.cars');
-    $container.insertAdjacentHTML('beforeend', loadingHTML);
-  });
-};
-
-export const removeLoadingHTML = () => {
-  $$('.loading').forEach($loading => $loading.remove());
-};
-
-export const renderProcess = carList => {
-  const moveHTML = `<div class="forward-icon mt-2">⬇️️</div>`;
-  const $cars = document.querySelectorAll('.result-container');
-
-  let index = 0;
-  const interval = setInterval(() => {
-    $cars.forEach($el => {
-      const carName = $el.querySelector('.car-player').innerHTML;
-      const html = carList.find(el => el.name === carName).process[index];
-
-      $el
-        .querySelector('.cars')
-        .insertAdjacentHTML('afterbegin', `${html === true ? moveHTML : ''}`);
-    });
-    index += 1;
-  }, 1000);
-
-  return interval;
-};
-
-export const timeoutSetting = (interval, trialCount) => {
-  setTimeout(() => {
-    clearInterval(interval);
-    alertCongrats();
-    removeLoadingHTML();
-    showWinner();
-  }, trialCount * 1000);
-};
-
 export const hideWinner = () => {
   $('.winner-section').classList.add('hide');
 };
@@ -121,4 +56,23 @@ export const hideWinner = () => {
 export const updateWinner = winners => {
   $('.winners').innerHTML = `
 	🏆 최종 우승자: ${winners.map(winner => `${winner}`).join(', ')} 🏆</h2>`;
+};
+
+export const resetName = () => {
+  resetNameForm();
+  toggleDisabledName();
+  focusNameInput();
+};
+
+export const resetTrial = () => {
+  resetTrialForm();
+  toggleDisabledTrial();
+  hideTrialForm();
+};
+
+export const resetResult = () => {
+  updateDistance([]);
+  hideDistance();
+  updateWinner([]);
+  hideWinner();
 };
